Rename `env-utils.ts` → `env-utils.client.ts`
